+++ super_gui/index.html
@@ -10,7 +10,7 @@
       @import url('https://fonts.googleapis.com/css2?family=Bebas+Neue&display=swap');
       </style>      
   </head>
-  <body>
+  <body class="bg-brown">
     <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
