+++ super_gui/src/App.tsx
@@ -4,17 +4,13 @@ import { StyleGuide } from "./pages/StyleGuide";
 import { Complaint } from "./components/complaint/Complaint";
 import Header from "./components/Header";
 import SmartLockUI from "./components/SmartLockUI";
+import { Nav } from "./global/Nav";
 
 const App = () => {
   return (
     <BrowserRouter>
       <Header />
-      <nav className=" flex justify-center gap-8">
-        {/* Navigation component */}
-        <Link to="/styleguide">Style Guide</Link>
-        <Link to="/smartlockui">Smart Lock UI</Link>
-        <Link to="/complaint">Complaint</Link>
-      </nav>
+      <Nav />
       <Routes>
         <Route path="/styleguide" element={<StyleGuide />} />
         <Route path="/smartlockui" element={<SmartLockUI />} />
