+++ super_gui/src/components/Header.tsx
@@ -2,8 +2,8 @@ import React from "react";
 
 const Header = () => {
   return (
-    <header className="my-4 text-center">
-      <h1 className="font-[Bebas_Neue] text-6xl text-[#D23715]">super</h1>
+    <header className="text-center border border-b-brown">
+      <h1 className="font-[Bebas_Neue] bg-white text-6xl text-[#D23715]">super</h1>
     </header>
   );
 };
