+++ super_gui/src/components/SmartLockUI.tsx
@@ -1,40 +1,12 @@
-import { Link } from "react-router-dom";
-import FrontDoor from "/assets/icons/bx-door-open.svg";
-import Parking from "/assets/icons/bxs-parking.svg";
-import Complaints from "/assets/icons/bxs-folder-open.svg";
-import SmartLocker from "/assets/icons/bxs-package.svg";
+
 import User from "/assets/icons/bxs-user.svg";
-import Lease from "/assets/icons/bxs-pen.svg";
 import Lock from "/assets/icons/bx-lock-alt.svg";
 
 const SmartLockUI = () => {
-  const menuItems = [
-    { Icon: FrontDoor, alt: "FrontDoor" },
-    { Icon: Parking, alt: "Parking" },
-    { Icon: Complaints, alt: "Complaints" },
-    { Icon: Lease, alt: "Lease" },
-    { Icon: SmartLocker, alt: "SmartLocker" },
-  ];
+  
 
   return (
     <main className="min-h-screen relative pb-16 md:pb-0">
-      {/* DESKTOP NAV BAR LAYOUT*/}
-      <div className="hidden md:block container mx-auto border-2 max-w-4xl">
-        <div className="border-b-2 border-t-2 p-2 flex justify-between">
-          {menuItems.map((item, index) => (
-            <div key={index} className="flex gap-2 items-center w-auto">
-              <Link
-                to=""
-                className="border rounded-full p-2 flex-shrink-0 bg-gray-800 hover:bg-accentGreen focus:bg-accentGreen focus:ring-0 focus:ring-accentGreen focus:outline-none transition duration-500 group"
-              >
-                <img src={item.Icon} alt={item.alt} className="w-6 h-6" />
-              </Link>
-              <p className="text-base font-medium">{item.alt}</p>
-            </div>
-          ))}
-        </div>
-      </div>
-
       {/* MAIN CONTENT */}
       <main className="mx-auto max-w-4xl m-4 p-4">
         <h1 className="text-center text-xl sm:text-2xl md:text-3xl font-semibold mb-4">
@@ -148,26 +120,6 @@ const SmartLockUI = () => {
           </div>
         </div>
       </main>
-
-      {/* MOBILE NAV BAR */}
-      <div className="md:hidden fixed bottom-0 left-0 right-0 border-t-2 p-2 bg-white shadow-lg">
-        <div className="flex justify-around max-w-4xl mx-auto">
-          {menuItems.map((item, index) => (
-            <div
-              key={index}
-              className="flex flex-col gap-1 items-center w-auto"
-            >
-              <Link
-                to=""
-                className="border rounded-full p-2 flex-shrink-0 bg-gray-800 hover:bg-accentGreen focus:bg-accentGreen focus:ring-0 focus:ring-accentGreen focus:outline-none transition duration-500 group"
-              >
-                <img src={item.Icon} alt={item.alt} className="w-5 h-5" />
-              </Link>
-              <span className="text-xs">{item.alt}</span>
-            </div>
-          ))}
-        </div>
-      </div>
     </main>
   );
 };
